Read initial routes on Android only
